+++ src/ButtonStyle.lua
@@ -1,0 +1,9 @@
+--!strict
+
+return {
+	Primary = 1,
+	Secondary = 2,
+	Success = 3,
+	Danger = 4,
+	Link = 5,
+}