+++ src/init.lua
@@ -1,5 +1,8 @@
 --!strict
 
+local ActionRowBuilder = require("./ActionRowBuilder")
+local ButtonBuilder = require("./ButtonBuilder")
+local ButtonStyle = require("./ButtonStyle")
 local Client = require("./Client")
 local EmbedBuilder = require("./EmbedBuilder")
 local GatewayOpcode = require("./GatewayOpcode")
@@ -12,6 +15,9 @@ local SessionStore = require("./SessionStore")
 local SlashCommandBuilder = require("./SlashCommandBuilder")
 
 return {
+	ActionRowBuilder = ActionRowBuilder,
+	ButtonBuilder = ButtonBuilder,
+	ButtonStyle = ButtonStyle,
 	Client = Client,
 	EmbedBuilder = EmbedBuilder,
 	GatewayOpcode = GatewayOpcode,